+++ tests/test_liquidity_universe.py
@@ -46,6 +46,15 @@ def test_grouped_liquidity(persistent_test_client: Client):
     assert max_liq == pytest.approx(31747.889)
     assert min_liq == pytest.approx(550)
 
+    # Test by sample by timestamp
+    amount, delay = liquidity_universe.get_liquidity_with_tolerance(
+        sushi_usdt.pair_id,
+        pd.Timestamp("2021-12-31"),
+        tolerance=pd.Timedelta("1y"),
+    )
+    assert amount == pytest.approx(2292.4517)
+    assert delay == pd.Timedelta('4 days 00:00:00')
+
 
 def test_combined_candles_and_liquidity(persistent_test_client: Client):
     """Candles and liquidity data looks similar regarding the pair count."""